feat(preprocessor): Implement UCN and wide char literal support (#599)
- Update `PPTokenKind::CharLiteral` and `TokenKind::CharacterConstant` to store `u64` instead of `u8`.
- Update `Literal::Char` to store `u64`.
- Implement `parse_ucn_escape` in `src/ast/literal_parsing.rs` to handle `\u` and `\U` escape sequences.
- Update `PPLexer::lex_char_literal` to use `literal_parsing::parse_char_literal` and `u64` type.
- Update `AstDumper` to format `LiteralChar` using unicode characters.
- Add regression test `src/tests/pp_wide_char.rs`.
- Verify all tests pass.

This enables correct evaluation of wide character constants (e.g. `L'\u0400'`) in preprocessor directives and supports UCNs in character literals.

+++ src/ast/literal.rs
@@ -22,6 +22,6 @@ pub enum FloatSuffix {
 pub enum Literal {
     Int { val: i64, suffix: Option<IntegerSuffix> },
     Float { val: f64, suffix: Option<FloatSuffix> },
-    Char(u8),
+    Char(u64),
     String(NameId),
 }
